requirements.txt
# requirements.txt
# AlphaBetting Working Requirements
# Installed on $(date)

# Core ETL
pandas>=2.0.0
psycopg2-binary>=2.9.6
requests>=2.31.0
pyyaml>=6.0
tenacity>=8.2.0
python-dotenv>=1.0.0

# Dependencies (auto-installed)
# numpy>=1.24.0
# python-dateutil>=2.8.0
# pytz>=2023.0
# tzdata>=2023.0

# To install later:
# scikit-learn>=1.3.0
# scipy>=1.10.0
# sqlalchemy>=2.0.0
# prefect>=2.10.0
# jupyter>=1.0.0
